Finish CCRFailureMapper move
