+++ lib/chef_apply/action/install_chef.rb
@@ -21,8 +21,7 @@ require "fileutils"
 
 module ChefApply
   module Action
-    class InstallChef < ChefApply::Action::Base
-
+    class InstallChef < Base
       def initialize(opts = { check_only: false })
         super
       end
@@ -54,13 +53,13 @@ module ChefApply
         # TODO BOOTSTRAP - we'll need to implement this for both platforms
         # require "mixlib/install"
         # installer = Mixlib::Install.new({
-        #   platform: "windows",
+        #   platform: "windows",/etc -
         #   product_name: "chef",
         #   channel: :stable,
         #   shell_type: :ps1,
         #   version: "13",
         # })
-        target_host.run_command! installer.install_command
+        # target_host.run_command! installer.install_command
         raise NotImplementedError
       end
